Rename the `Base` class file
